+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Replaced `express-graphql` with `graphql-http` - [#31](https://github.com/superfaceai/one-service/pull/31)
+- Moved GraphiQL from `/graphql` to root `/` - [#31](https://github.com/superfaceai/one-service/pull/31)
 
 ## [2.0.4] - 2023-02-16
 ### Changed